# services/spring-order/generated/src/main/java/org/openapitools/api/ApiApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.22.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.AuthResponse;
import org.openapitools.model.LoginRequest;
import org.openapitools.model.RegisterRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-05-12T13:18:18.645670241+02:00[Europe/Berlin]", comments = "Generator version: 7.22.0")
@Validated
@Tag(name = "api", description = "the api API")
public interface ApiApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_API_V1_AUTH_LOGIN_POST = "/api/v1/auth/login";
    /**
     * POST /api/v1/auth/login : Login user
     *
     * @param loginRequest  (required)
     * @return Login successful (status code 200)
     */
    @Operation(
        operationId = "apiV1AuthLoginPost",
        summary = "Login user",
        responses = {
            @ApiResponse(responseCode = "200", description = "Login successful", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AuthResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ApiApi.PATH_API_V1_AUTH_LOGIN_POST,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<AuthResponse> apiV1AuthLoginPost(
        @Parameter(name = "LoginRequest", description = "", required = true) @Valid @RequestBody LoginRequest loginRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"token\" : \"token\", \"username\" : \"username\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_API_V1_AUTH_REGISTER_POST = "/api/v1/auth/register";
    /**
     * POST /api/v1/auth/register : Register new user
     *
     * @param registerRequest  (required)
     * @return User registered successfully (status code 200)
     */
    @Operation(
        operationId = "apiV1AuthRegisterPost",
        summary = "Register new user",
        responses = {
            @ApiResponse(responseCode = "200", description = "User registered successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AuthResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ApiApi.PATH_API_V1_AUTH_REGISTER_POST,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<AuthResponse> apiV1AuthRegisterPost(
        @Parameter(name = "RegisterRequest", description = "", required = true) @Valid @RequestBody RegisterRequest registerRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"token\" : \"token\", \"username\" : \"username\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}